backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
face_recognition==1.3.0
sentence-transformers==3.0.0
neo4j==5.21.0
razorpay==1.4.0
python-dotenv==1.0.1
google-generativeai==0.7.2
Pillow==10.3.0
numpy==1.26.4
opencv-python==4.10.0.82
pyzbar==0.1.9
requests==2.32.3
aiofiles==23.2.1
websockets==12.0
